refactor: decompose 901-line kinematic_forces.py into domain modules (fixes #2386)
